+++ etl/pad/storage_processor/series_processor.py
@@ -5,7 +5,7 @@ from collections import defaultdict
 
 from pad.db.db_util import DbWrapper
 from pad.raw_processor import crossed_data
-from pad.storage.monster import MonsterWithSeries
+from pad.storage.series import MonsterSeries
 from pad.storage.series import Series
 
 __location__ = os.path.realpath(os.path.join(os.getcwd(), os.path.dirname(__file__)))
@@ -30,9 +30,12 @@ class SeriesProcessor(object):
 
     def _try_ancestor(self, db: DbWrapper):
         # For monsters with no series, try applying the series of the ancestor.
-        monster_id_to_series_id = db.load_to_key_value('monster_id', 'series_id', 'monsters')
+        monster_id_to_series_id = db.load_to_key_value('monster_id', 'series_id', 'monster_series', 'priority = 1')
         monster_id_to_ancestor_id = db.load_to_key_value('to_id', 'from_id', 'evolutions')
-        for monster_id, series_id in monster_id_to_series_id.items():
+        for csc in self.data.ownable_cards:
+            monster_id = csc.monster_id
+            series_id = monster_id_to_series_id.get(monster_id, 0)
+
             if series_id != 0:
                 # Series already set.
                 continue
@@ -48,22 +51,18 @@ class SeriesProcessor(object):
                 continue
 
             # Apply the ancestor series to the current monster
-            item = MonsterWithSeries(monster_id=monster_id, series_id=ancestor_series_id)
+            item = MonsterSeries(monster_id=monster_id, series_id=ancestor_series_id, priority=True)
             db.insert_or_update(item)
 
     def _try_group(self, db: DbWrapper):
         # Try to infer the series of a monster via the series of monsters in its group.
-        monster_id_to_series_id = db.load_to_key_value('monster_id', 'series_id', 'monsters')
+        monster_id_to_series_id = db.load_to_key_value('monster_id', 'series_id', 'monster_series', 'priority = 1')
 
         # Partition existing DadGuide monster series by PAD group ID.
         group_id_to_series_ids = defaultdict(set)
         for csc in self.data.ownable_cards:
             monster_id = csc.monster_id
-            series_id = monster_id_to_series_id.get(monster_id)
-            if series_id is None:
-                monster_id_to_series_id[monster_id] = 0
-                logger.warning('Series was null for monster %d', monster_id)
-                continue
+            series_id = monster_id_to_series_id.get(monster_id, 0)
             if series_id == 0:
                 # No useful data from this card
                 continue
@@ -74,7 +73,7 @@ class SeriesProcessor(object):
         # Now loop through again and see if we can apply any series via group mapping.
         for csc in self.data.ownable_cards:
             monster_id = csc.monster_id
-            series_id = monster_id_to_series_id[monster_id]
+            series_id = monster_id_to_series_id.get(monster_id, 0)
             if series_id != 0:
                 # Series already set.
                 continue
@@ -83,5 +82,9 @@ class SeriesProcessor(object):
             possible_series = group_id_to_series_ids[group_id]
             if len(possible_series) == 1:
                 group_series_id = list(possible_series)[0]
-                item = MonsterWithSeries(monster_id=monster_id, series_id=group_series_id)
+                item = MonsterSeries(monster_id=monster_id, series_id=group_series_id, priority=True)
+                db.insert_or_update(item)
+            else:
+                # Give up and insert with series_id 0
+                item = MonsterSeries(monster_id=monster_id, series_id=0, priority=True)
                 db.insert_or_update(item)
